feat: generate rust lib inside this repo
the problem here was that the generated code couldn't get picked up by
the backend so we can "generate" that code into lib.rs and export that
lib to be imported to our other crates. This will allow us to use the
other code in other crates. We can also drop the dependency on
tonic-build in other crates because of this